// src/types/folders.ts
export type Folders =
  | 'achievements'
  | 'artifacts'
  | 'bait'
  | 'character_exp_material'
  | 'characters'
  | 'common_materials'
  | 'domains'
  | 'elemental_stone_materials'
  | 'fish'
  | 'fishing_rod'
  | 'furnishing'
  | 'food'
  | 'geography'
  | 'ingredients'
  | 'jewels_materials'
  | 'local_materials'
  | 'monsters'
  | 'potions'
  | 'talent_lvl_up_materials'
  | 'tcg_action'
  | 'tcg_characters'
  | 'tcg_monsters'
  | 'weapon_enhancement_material'
  | 'weapon_primary_materials'
  | 'weapon_secondary_materials'
  | 'weapons';